backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
python-docx==1.1.0
pypdf2==3.0.1
markdown==3.5.2
aiofiles==23.2.1
pydantic==2.5.3
celery[redis]==5.3.6
redis==5.0.1
boto3==1.34.34
